econ-dashboard/requirements.txt
streamlit==1.32.0
plotly==5.19.0
pandas==2.2.0
google-cloud-bigquery==3.17.0
db-dtypes==1.2.0
requests==2.31.0
